Abstracting D3 tree representation as RrbVec class
