+++ web-vis/www/index.html
@@ -11,5 +11,7 @@
       <button type="button" id="vectorSplitter">Split vector in the middle</button>
     </div>  
     <script src="./bootstrap.js"></script>
+    <div id="tree1"/>
+    <div id="tree2"/>
   </body>
 </html>
